This is a massive update.
I have updated most all files to handle the move from /maps to the site root /.  This means that nearly all HREF's have been replaced with the server root notation.  This was wreaking havoc when the master template was loaded into a folder other than the root.

Folders have been created along with their respective index files for the main parts of the site.  Some fairly crude placeholders are in effect for now.  Links on the master template have been updated and correspond to correct locations.

The user class was updated and is now fully working.  The main issue was cookies, and those have been sorted out.

The master template now checks whether or not the user is authenticated and reacts accordingly.

Major files have been commented so that they can be qualified from more than file name.

The main page is now a generic home page instead of the route creation page.  The route creation page has been moved to its correct home at routes/create.php.

Finally, previous templates have been indented correctly.  Any remaining ones that look bad are probably about to be deleted (this really applies to the create page which has been corrected).

I think we are almost ready to push these to the main server and call it an alpha release!