Includes my major changes for the night.  There was some overlap between things we did tonight, and I will detail those in a note.
First, you accidentally deleted the definition for black from image_route.php.  This caused an error on my pages, I added that value back in.

The viewing route template now has a section which shows up on routes that you are the actual creator.  Right now there is dummy content, but that will soon hold a toolbox of sorts for deleting or editing routes.

The route index page has the most signifcant changes.  The template now references a route array in the current user.  The signficance of that is that the routes now belong just to the user.  There was not much else to change from what you did, luckily no other variables had their name changed.  The index.php page for routes looks a little different too.  I changed the function which loads the data into the current user.

The user class now has a function to load a given number of routes into an array.  This is used on dashboard pages for showing so many routes.  Added in a helper function to generate a user from a given database query.  This is similar to the fromFetchAssoc in the Routes class.  The main difference is that there is an instance and a static implementation (look at where they are used to see why).  Removed some random comments from this class as well.  Removed some redundant code and replaced it with the new loadFromFetchAssoc function.

The community index page now has some content.  The goal here was to list users and link to valid profile pages.  These are really placeholders.  The profile pages will actually be filled with content soon.

The route class now has a variable for the user id.  This is used on the view page to check if the current user is the route creator.