Updated the main javascript file to allow for generic objects for storing settings (map and user).
Changed the route and settings form pages.

Updated the user class to add the home location to the settings grabbed in fromFetchAssoc.

Updated the settings and create templates so that they start on the user's location if it exists.

View route template now uses the same styling as the create template (mile markers and such