Did a bit of work on the user class to clean up the way that users are validated.  The main idea is to return users instead of doing a lot of stuff behind the scenes inside the classes.
The classes no longer hold a private reference to the database.  The issue was that this was being held in the session data, becoming invalidated, and then throwing errors.

The route class was updated.  Removed the round distance function.

Updated several templates to remove the dots in the title.

The settings template was updated to included similar settings as the register page.

The config class was split into two separate files.  The first file contains the class references for necessary classes (user, route, database).  The second config file is named the same and contains the details for the templating engine.