Added in most of the feature that was needed.
A page exists to add a friend via ajax from the user's profile page.  Added a list of user's to the index page of the community page.  Updated the user class to include a getFriends function.  Also changed the return of the cookieLogin function to avoid a non-User return from the function.