Includes a large amount of code related to creating a general purpose calendar.  This calendar is sent to Smarty for rendering.  The calendar has the ability to hold arbitrary data which is then rendered with the correct Smarty tags.
Anyways, with that in place I intend to extend the training class to get the events that are within the given range.  The calendar has a bit of potential on this page but I admit that it has its drawbacks.  The main one is that I am using a list which requires floating.  Smarty does not provide an elegant way to do this with tables.  Let me know what you think of the calendar in general as a way of displaying this data, and then specifically how I have coded it.

See #88.