See #100.
This adds in the code to allow for user activation.  The user cookie hash is now set at user creation and remains constant.  That hash is used to activate the user.  The activation goes through the action_login.php?action=activate page.  There will be an email sent on user creation, but that cannot be done quite yet.  Also cleaned up parts of the user class so that the coordination of the parts is cleaner.