See #124.
This changes the layout of the create route page.  It uses the collapsable side panel like we had discussed.  It adds in the JQuery dimension plugin for the resizing methods.  There is a bit of javascript to change the height and toggle the side panel.  Also removed a stray line from the master template while I was in there.