See #99.  See #124.
This integrates branch 124 into branch 99.  This was needed so that the layout was correct.  The changes to 99 include adding the ability to edit and delete routes.  There are also significant changes to the route class to make this process easier.  There were also changes to the action page for the routes using all of the new features to simplify the process.

The delete feature is handled using Thickbox.  The files for that were added back in.  The master template was changed to add in these files.  The stylesheets should be combined eventually.

With all of this in place, I will now look into the second set of requirements which are that routes should hold references to each other after edited.