See #99.
Did more work on editing and deleting routes.  The changes now allow for a hierarchy of routes.  It allows for a route to be edited if allowed or copied if not allowed to edit.  Worked on the javascript for the map so that things work better (or at all).  This is almost done and I will post the SQL alter script later.