See #88.
Did a bit of work on the training class.  Combined the TraingingLogItem into the TrainingLog class.  Also moved toward using the new method of CRUD operations.  Renamed the action page.  Added documentation to the training class and the calendar class.

I believe this is ready to move into the trunk and face scrunity in new tickets.