See #88.
This integrates branch 88 back into the trunk.  This is a fairly major branch that is being brought back in.  Most of the files are new and there are really only slight changes to the other files that matter.  Anyways go through the history of this ticket to see all the changes that have been added.  This will probably close the ticket, but I will do that manually.