See #99.
This integrates branch 99 back into the trunk.  This also brings in branch 124 since it was moved into 99 and has not changed since.  Both of those tickets can probably be closed but that will be done manually.  Also the SQL alter script will be posted.

The main changes here were to the route templates include create and view.  It allows for a route to be edited or deleted based on various conditions.  There is more to be done, but I want to get this into the main trunk for now.