Closes #119.
Integrates branch 119 back into the trunk.  This moves the external javascript calls to the pages where they are used.  This removes the mapping javascript from the master template and puts it into the view and create templates where it is currently used.

The main reason for this is to increase load times on teh pages that do not require javascript.