See #97.
Adds a validation plugin to the site.  The plugin uses JQuery and simplifies the validaiton process.  This plugin was them implemented on all the existing forms.

There was an addition to the user class to check if a username exists.  This was added for AJAX checking of duplicate user names.

There were changes to the way the training class handles its date property.  I added a call to strtotime in the fromFetchAssoc function.  This caused a couple of changes on the training pages related to accessing that date.  This was really needed to make a standard way of handling the date (which is now a unix timestamp).

Delete an action page which is no longer needed.  Added saving support to the training class and the action_trianing page.