Closes #97.
Integrates branch 97 back into the trunk.  This adds support for form validation.  This is done by a JQuery plugin.  Seems to be a pretty easy way to do validation that is standard across all of our pages.  This creates a little bit of styling issues, but we can sort those out.