See #126.
Adds in the core files for viewing help pages.  The main idea is that help pages will be loaded into Thickbox from the page that needs the info.  This pages are accessed by a common name and a simple view.php inside the help folder.  The templates are unique for each help page and they are loaded into a mini-master template for the help files.  Requires a quick change to the database for the common name field (p_common).