See #139.
Adds in the google tracking code.  Currently this adds the code for "byroni.us" and its subdomains.  If we do want to use this feature then we will have to change it when we go the real site.

Using Firebug to see the network requirements, it adds anywhere from 40-100ms to the connection.  The javascript that it loads is cached which helps after the first time.  This seems reasonable.  Even on my connection where it is going to the localhost, the page still requires 290ms total for the community page (which is the most simple).