See #143. See #98.
This does a little more updating to the browse page.  Also this adds in the code for creating custom SQL queries from form input.  Those classes are not in the right place quite yet.  That is the final thing that I have to change around.  I also went ahead and fixed a couple of errors in the javascript related to viewing in IE.  Not the best place to do this, but I saw them earlier while using IE and wanted to fix them.