Closes #126.
Integrates branch 126 back into the trunk.  The help feature is by no means complete.  It needs to be brought in however so that we can finish it up in terms of the overall design.  The basic idea is to have all the help pages with their own template which can be brought in in a Thickbox or on an actual help page.