Closes #135.
Integrates branch 135 back into the trunk.  This adds RSS feeds for training and activity data.  These are not complete, but the remaining issues are a matter of content.  I want this to be in the trunk now so we have an idea of all the features that do and do not exist.