Closes #143.
Integrates branch 143 back into the trunk.

This is a fairly massive integration which affected a whole lot more than originally planned.

It extends the Smarty class to add functions which will be useful.  The notable one here is display_master($page_name) which loads a template into the master template.  Config files were update for this class and also the new SQL parser.

The info classes were updated to get a result set that could be paged.  This was quickly redone by the SQL parser but the original function still exists because it is called once.  This will be removed soon by a general getResults() function.

Browse pages were added for the routes and training items.  This supports the listing on the browse.php page but also allows for querying from other pages.  This was the major addition for this branch.

The footer and master templates were updated to add links and also the JQuery form plugin which allows for simple AJAX submit of forms (not used yet) and also resetting and some other functions (used).

Finally, the *_list.tpl files were added to their respective locations.  This is the first in the push to move toward modular design.  Those files are intended to take an array of objects and render generic, uniform HTML.