See #152.
This is a good first prototype for the module system.  I have added a field to the user table to test the preferences part.  Take a look at the changes as they are the important things I added.

I also put two really useful functions in the config class file.  The first safely grabs the value from an array (checks isset).  The second turns an array into a class.  It does the job of our fromFetchAssoc functions but without us defining all the fields.  I think that is very cool.  Anyways take a look.