See #157.
This adds in the admin tab.  The admin tab currently allows for updating the site stats, viewing the site stats, and managing the pages.  The page thing is a table of forms which allow for changing the page properties.  They are sent via AJAX to the server so that updating them is simple and quick.  I will add in user management and brind this into the trunk.