See #157.
This adds support for getting users.  It is a crude implementation but it will filter out users.  The update actions do not do anything yet.  I am not sure what we want to be able to do from here.  Also the user class is about to get re-done so I am waiting to do anything solid here.  Regardless, I am going to bring this into the trunk for the admin part and the pages.