Closes #81. Closes #141. Closes #148. Closes #153. Closes #154. Closes #158. See #128.
This is an absolutely comprehensive cleanup of the source. I have removed the config_class file.  I have removed all of the require_once statements.  They were replaced with the _autoload.  Doing that required me to name a few files differently.

I have replaced every single reference to fromFetchAssoc to the new function.  This required me to extend all of the classes with the Object class.

All of the classes now exactly match the database in terms of variables names.  This excludes the prefix.  The prefix is not a big deal because the new array to class function accounts for it.

The User class has been slightly cleaned up.  There is now a static reference to currrent user after the config page.  This means that _SESSION[userData] only appears on the config page and once in the user class.  There is a little work remaining on the user class.

Every instance of (header: location...) was replaced by the new Page::redirect function.

I have told all the php files to use the new display_master function.