See #156.
This adds in feedback from the user.  This is currently given a tab, but that will be moved into the header bar once styled.  This works for logged in and not logged in users.  I also created the admin page which we can use to look over this stuff.  The admin page can delete messages.