See #103.
Adds a create form for training entries.  This form knows which routes and training types are available from the database.  There is a little javascript to manipulate whether or not the user is recording this for a known route or not.  Currently only works for routes associated with the current user.