See #163.
This changes about every file.  The main changes in every file are to remove the require("config") statement, and the $smarty->display call.  These are now put into the main index file which all calls now go through.  This greatly reduces code and confusion and provides a central place to get functions included along with settings and constants.  There are a couple more things which need to be done.  These include specific actions for calls which need to return via AJAX or RSS.

There is probably a little more housekeeping.  This includes deleting all the files which now just have <?php ?> in them.