See #165.
This adds in the modules.  The database has been changed slightly so update in order to avoid errors (added a users_modules table to unclutter the main users table).  The modules are in place on all of the index pages.  Most of the added files are templates.  I also changed all the remaining bad links to correct ones.