See #235.
This adds elevation data to the view routes page.  This is done by storing elevation data in the database.  If the data is available after a route is created, then it is retrieved and the elevation profile is stored in the routes table.  This data is then pulled by the view template and presented with a Flot chart.

NOTE: This introduces a breaking change to the code that produces route images.  This was done to allow for the decoding of the polyline in a manner that is easier to process for elevation purposes.  This will need to be fixed when we re-introduce route images.