See #236.
This includes the code to allow for URLs to contain parameters without using the query string.  Any controller action that wants to can name parameters using the RoutingEngine::getInstnace()->registerParams() function.  That function is called with a list of names that should be associated with the parameters in order.  These values are then stored in the _GET array.  This allows old code to work while still using a better looking URL.  I have also updated the major controller actions that are affected by this.  Finally, I have updated any links to use the new format.