See #241.
This fixes the two issues on this page.  The first was the javascript source error described earlier.  The other error is apparently using onkeyup:true for the validator.  After searching around it appears that this is not allowed.  We can either set it to false or assign a function, but not to "true".  I will update the server and will also fix the onkeyup error in one of my branches so that it propagates.