See #187.
This adds in the code to undo entire operations instead of single points.  This works for out and back and follow roads.  The mechanism is very simple and limited.  If we need a more robust undo solution, it is possible, but this works for now.

It simply keeps track of the number of points that need to be removed the next time an undo is called.  The undo function retrieves this number of and then removes that many points.