See #239.
The added code here gets rid of the old database call for the page details.  The details are now supplied in the source code.  This provides a way to self document the controller actions.  If there is not a call to the setPage function then the details are filled in from defaults (title: runnDAILY, perm: ADMIN).  I think the effort of adding this code to every controller action is ultimately more sustainable than a datbase table full of page titles which has to be updated whenever a controller or action is added.

The home controller has a sample of how to use the new call.  Finally, the Facebox changes are something I broke last time.