See #253.
I realized I am working in the wrong branch.  Anyways, this extends the SQL class to add more features.  This allows for array vs. object return.  Also, for object return if there was a left-join, it can take that data and put it into the main object.  See routes/view for an example.