See #253.
This is a mostly complete version of the SQL additions.  I have gone through and replaced a lot of the controller pages with calls to these methods instead.  I am not sure if it should be the permanent solution or not.  I am also not sure about any performance hit we take.  I will look into those things more.  I will also post some code to Trac about how to use the class.

If this doesn't break anything, I would like to get it into the trunk soon, so we can at least try it out on while we develop.