See #254.
This includes a nearly complete version of the new templating engine.  Right now it is ignoring modifiers, but I will add those next.  I will also begin to document how this works so that we can both be fully aware of what is happening.  I always felt like Smarty was a guessing game what it would do, so I want to make sure we both understand this one.  As usual, I am not completely finished, so if you want to try it, let me know what you get.  The main idea is that it loads all of the include calls into a single template file so that run time is much faster.  Right now it is faster than Smarty and compiles the templates fresh every time.  Turning that off, I was getting around 1ms for static pages and 4ms for the others.