See #254.
This brings the templating engine one step closer to completion.  We now have support for functions, blocks, and modifiers.  The functions and modifiers are both copied into the compiled file directly.  Now is just a matter of catching errors on pages to find the things I missed.  Also porting functions from Smarty is painless and nearly just copy and paste.  We could batch script all of the functions and modifiers to be fully compliant with Smarty, or just use the ones we need.  Also changes a couple templates which were doing odd things for no reason.